+++ contracts/mixins/Constants.sol
@@ -2,39 +2,36 @@
 
 pragma solidity ^0.8.0;
 
+/// Constant values shared across mixins.
+
 /**
- * @title Constant values shared across mixins.
+ * @dev 100% in basis points.
  */
-abstract contract Constants {
-  /**
-   * @notice 100% in basis points.
-   */
-  uint256 internal constant BASIS_POINTS = 10000;
+uint256 constant BASIS_POINTS = 10000;
 
-  /**
-   * @notice Cap the number of royalty recipients.
-   * @dev A cap is required to ensure gas costs are not too high when a sale is settled.
-   */
-  uint256 internal constant MAX_ROYALTY_RECIPIENTS = 5;
+/**
+ * @dev Cap the number of royalty recipients.
+ * A cap is required to ensure gas costs are not too high when a sale is settled.
+ */
+uint256 constant MAX_ROYALTY_RECIPIENTS = 5;
 
-  /**
-   * @notice The minimum increase of 10% required when making an offer or placing a bid.
-   */
-  uint256 internal constant MIN_PERCENT_INCREMENT_DENOMINATOR = BASIS_POINTS / 1000;
+/**
+ * @dev The minimum increase of 10% required when making an offer or placing a bid.
+ */
+uint256 constant MIN_PERCENT_INCREMENT_DENOMINATOR = BASIS_POINTS / 1000;
 
-  /**
-   * @notice The gas limit used when making external read-only calls.
-   * @dev This helps to ensure that external calls does not prevent the market from executing.
-   */
-  uint256 internal constant READ_ONLY_GAS_LIMIT = 40000;
+/**
+ * @dev The gas limit used when making external read-only calls.
+ * This helps to ensure that external calls does not prevent the market from executing.
+ */
+uint256 constant READ_ONLY_GAS_LIMIT = 40000;
 
-  /**
-   * @notice The gas limit to send ETH to multiple recipients, enough for a 5-way split.
-   */
-  uint256 internal constant SEND_VALUE_GAS_LIMIT_MULTIPLE_RECIPIENTS = 210000;
+/**
+ * @dev The gas limit to send ETH to multiple recipients, enough for a 5-way split.
+ */
+uint256 constant SEND_VALUE_GAS_LIMIT_MULTIPLE_RECIPIENTS = 210000;
 
-  /**
-   * @notice The gas limit to send ETH to a single recipient, enough for a contract with a simple receiver.
-   */
-  uint256 internal constant SEND_VALUE_GAS_LIMIT_SINGLE_RECIPIENT = 20000;
-}
+/**
+ * @dev The gas limit to send ETH to a single recipient, enough for a contract with a simple receiver.
+ */
+uint256 constant SEND_VALUE_GAS_LIMIT_SINGLE_RECIPIENT = 20000;