+++ contracts/mixins/NFTMarketCore.sol
@@ -18,7 +18,7 @@ error NFTMarketCore_Seller_Not_Found();
  * @title A place for common modifiers and functions used by various NFTMarket mixins, if any.
  * @dev This also leaves a gap which can be used to add a new mixin to the top of the inheritance tree.
  */
-abstract contract NFTMarketCore is Constants, Initializable {
+abstract contract NFTMarketCore is Initializable {
   using AddressUpgradeable for address;
 
   /// @notice The FETH ERC-20 token for managing escrow and lockup.